Date and Categories Update
Date Buttons were added to check for all rows concerning "Today", "Tomorrow", "This Week", and "Past Due." The format to enter dates will be: YYYY-MM-DD.

Categories were added to the interface, you are able to add categories into a list, select which category to filter the database from, delete categories from the listbox, and finally remove the category filter.
Note: All of the categories are saved in a SQLite Table and will remain between sessions.